requirements.txt
os
python-dotenv
tensorflow
numpy
math
shutil
